Preserve iy when executing a test
This avoids tests which destroy iy from bringing down the test runner,
because the pointer to the display buffer would be garbage.

This is done by putting iy on the stack before each test starts, and
popping it off again in the pass/fail routine.

Additionally, this commit makes the printDebugAndHalt routine a little
nicer: it also prints bc now, and it adds labels to indicate which
register is which in the output.